chore: add release workflow and unify binaries
