+++ .github/workflows/release.yml
@@ -1,17 +1,14 @@
-name: build
+name: release
 
 on:
   push:
-    branches:
-      - main
+    tags:
+      - 'v*'
   workflow_dispatch:
 
 jobs:
-  verify:
-    uses: ./.github/workflows/verify.yml
-  publish-latest:
+  release-predator:
     runs-on: ubuntu-latest
-    needs: verify
     steps:
       - name: Checkout
         uses: actions/checkout@v2
@@ -27,5 +24,11 @@ jobs:
           registry: docker.io
           username: ${{ secrets.DOCKERHUB_USERNAME }}
           password: ${{ secrets.DOCKERHUB_TOKEN }}
-      - name: Push latest predator image
-        run: env GOOS=linux GOARCH=amd64 make build && docker build -t docker.io/odpf/predator:latest . && docker push docker.io/odpf/predator:latest
+      - name: Run GoReleaser
+        uses: goreleaser/goreleaser-action@v2.6.1
+        with:
+          distribution: goreleaser
+          version: v1.8.3
+          args: --rm-dist
+        env:
+          GITHUB_TOKEN: ${{ secrets.GO_RELEASER_TOKEN }}
