+++ .goreleaser.yml
@@ -1,0 +1,59 @@
+project_name: predator
+
+before:
+  hooks:
+    - make install
+builds:
+  - main: ./cmd/main.go
+    id: "predator"
+    binary: predator
+    flags:
+      - -a
+    ldflags:
+      - -s -w -X github.com/odpf/predator/conf.BuildVersion={{.Tag}} -X github.com/odpf/predator/conf.BuildCommit={{.FullCommit}}
+    goos:
+      - linux
+      - darwin
+      - windows
+    goarch:
+      - amd64
+      - 386
+      - arm
+      - arm64
+    env:
+      - CGO_ENABLED=0
+archives:
+  - replacements:
+      darwin: macos
+      linux: linux
+      windows: windows
+      386: i386
+      amd64: x86_64
+    format_overrides:
+      - goos: windows
+        format: zip
+release:
+  draft: true
+  prerelease: auto
+checksum:
+  name_template: 'checksums.txt'
+snapshot:
+  name_template: "{{ .Tag }}-next"
+changelog:
+  sort: asc
+  filters:
+    exclude:
+      - '^docs'
+      - '^test'
+      - '^build'
+      - '^chore'
+dockers:
+  -
+    goos: linux
+    goarch: amd64
+    ids:
+      - predator
+    dockerfile: Dockerfile
+    image_templates:
+      - 'docker.io/odpf/{{.ProjectName}}:{{ .Version }}'
+      - 'docker.io/odpf/{{.ProjectName}}:{{ .Version }}-amd64'
