+++ conf/build.go
@@ -1,0 +1,9 @@
+package conf
+
+var (
+	// BuildVersion will be overridden by the build system
+	BuildVersion = "dev"
+
+	// BuildCommit will be overridden by the build system
+	BuildCommit = ""
+)
